Add files via upload
This is a Texas Hold'Em Hand Generator. This python file can be run to generate a shuffled deck of 52 cards that are dealt to two players on a "table", then generates a "flop", "turn",  and "river". After the shuffled deck is generated and dealt, the script calculates the winner of the two players (1:  Hero, 2: Villain) by analyzing what type of hand each player has. The possible hands are [ranked worst to best], High Card, Pair, Two Pair, Three of a Kind, Straight, Flush, Full House, Straight Flush, and Royal Flush. After the players hands are analyzed, the Hero Hand [player 1] is given a Win/Lose/Tie value 1:Win, -1:Lose,  0:Tie.